refactor(eagle3): hoist data/loss helpers into canonical components dirs
Two more PR #2224 review comments asked for these EAGLE-3 utilities to
live under the cross-cutting component directories instead of being
buried inside ``components/speculative/eagle/``:

- ``components/speculative/eagle/loss.py`` ->
  ``components/loss/soft_ce.py`` -- ``masked_soft_cross_entropy`` is a
  generic soft-target / KD-style CE that any future distillation
  recipe could reuse; it does not depend on anything EAGLE-3-specific.
- ``components/speculative/eagle/data.py`` ->
  ``components/datasets/llm/eagle3.py`` -- ``build_eagle3_dataloader``
  is a thin ``ChatDataset`` wrapper and ``build_eagle3_token_mapping``
  builds the draft sub-vocab; both fit the ``datasets/llm`` namespace
  where every other LLM dataset helper lives.

Update the three import sites (``eagle/core.py``,
``recipes/llm/train_eagle3.py``, ``tests/unit_tests/speculative/
test_eagle3.py``) to pull from the canonical paths directly rather
than going through an ``eagle.__init__`` re-export, so the
architectural intent is clear from the import line itself. The
``eagle/__init__`` re-exports for these helpers are dropped; only the
EAGLE-3-native classes (``Eagle3TrainerModule``,
``LlamaEagle3DraftModel``, ``HFEagle3TargetModel``) remain.
