+++ nemo_automodel/components/speculative/eagle/core.py
@@ -21,7 +21,7 @@ from dataclasses import dataclass
 import torch
 import torch.nn as nn
 
-from nemo_automodel.components.speculative.eagle.loss import masked_soft_cross_entropy
+from nemo_automodel.components.loss.soft_ce import masked_soft_cross_entropy
 
 
 def _shift_left_with_zero(tensor: torch.Tensor) -> torch.Tensor:
